Fix: Correct bar chart data mapping and improve formatting
This commit fixes a data mapping issue in the bar chart implementation, where the X and Y axes were swapped, and improves the overall chart presentation.

- **Data Mapping Fix:**
    - Corrected the `ChartData` model to properly map transaction amounts to the Y-axis and time to the X-axis. Previously, `amount` was on the X-axis and `time` on the Y-axis.
    - Updated `YesterdayChart` to correctly group datasets, assigning time-of-day (in seconds) to the `x` value and `amount` to the `y` value.

- **Chart Component (`VicoBarChart`):**
    - The chart now maps the original X-axis values (time in seconds) to indices for rendering, ensuring the X-axis labels correctly format back to a `HH:mm` time string.
    - The marker formatter has been updated to correctly display the original amount (`y`) and time (`x`) when a bar is highlighted.
    - The bar thickness has been increased from `1.dp` to `24.dp` for better visual representation.

+++ app/src/main/java/com/example/moneytracker/ui/components/charts/collections/ChartData.kt
@@ -3,8 +3,8 @@ package com.example.moneytracker.ui.components.charts.collections
 import androidx.compose.ui.graphics.Color
 
 data class ChartData(
-    val x: List<Int> = listOf(),
-    val y: List<Double> = listOf(),
+    val x: List<Double> = listOf(),
+    val y: List<Int> = listOf(),
     val color: Color = Color.Gray,
     // Optional label shown in legends. Keep default null to avoid breaking existing callers.
     val label: String? = null
